# Compute OTel metrics