--- a/common/otel_metrics.go
+++ b/common/otel_metrics.go
@@ -48,18 +48,23 @@ type otelMetrics struct {
 func (o *otelMetrics) GCSReadBytesCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.gcsReadBytesCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) GCSReaderCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.gcsReaderCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) GCSRequestCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.gcsRequestCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) GCSRequestLatency(ctx context.Context, value float64, attrs []MetricAttr) {
 	o.gcsRequestLatency.Record(ctx, value, attrsToRecordOption(attrs)...)
 }
+
 func (o *otelMetrics) GCSReadCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.gcsReadCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) GCSDownloadBytesCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.gcsDownloadBytesCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
@@ -67,9 +72,11 @@ func (o *otelMetrics) GCSDownloadBytesCount(ctx context.Context, inc int64, attr
 func (o *otelMetrics) OpsCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.fsOpsCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) OpsLatency(ctx context.Context, value float64, attrs []MetricAttr) {
 	o.fsOpsLatency.Record(ctx, value, attrsToRecordOption(attrs)...)
 }
+
 func (o *otelMetrics) OpsErrorCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.fsOpsErrorCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
@@ -77,9 +84,11 @@ func (o *otelMetrics) OpsErrorCount(ctx context.Context, inc int64, attrs []Metr
 func (o *otelMetrics) FileCacheReadCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.fileCacheReadCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) FileCacheReadBytesCount(ctx context.Context, inc int64, attrs []MetricAttr) {
 	o.fileCacheReadBytesCount.Add(ctx, inc, attrsToAddOption(attrs)...)
 }
+
 func (o *otelMetrics) FileCacheReadLatency(ctx context.Context, value float64, attrs []MetricAttr) {
 	o.fileCacheReadLatency.Record(ctx, value, attrsToRecordOption(attrs)...)
 }
@@ -126,5 +135,4 @@ func NewOTelMetrics() (MetricHandle, error) {
 		fileCacheReadBytesCount: fileCacheReadBytesCount,
 		fileCacheReadLatency:    fileCacheReadLatency,
 	}, nil
-
 }