--- a/common/telemetry.go
+++ b/common/telemetry.go
@@ -24,6 +24,8 @@ import (
 
 type ShutdownFn func(ctx context.Context) error
 
+// The default time buckets for latency metrics.
+// The unit can however change for different units i.e. for one metric the unit could be microseconds and for another it could be milliseconds.
 var defaultLatencyDistribution = metric.WithExplicitBucketBoundaries(1, 2, 3, 4, 5, 6, 8, 10, 13, 16, 20, 25, 30, 40, 50, 65, 80, 100, 130, 160, 200, 250, 300, 400, 500, 650, 800, 1000, 2000, 5000, 10000, 20000, 50000, 100000)
 
 // JoinShutdownFunc combines the provided shutdown functions into a single function.