--- a/tools/integration_tests/run_e2e_tests.sh
+++ b/tools/integration_tests/run_e2e_tests.sh
@@ -82,6 +82,7 @@ TEST_DIR_PARALLEL=(
   "concurrent_operations"
   "benchmarking"
   "mount_timeout"
+  "monitoring"
 )
 
 # These tests never become parallel as it is changing bucket permissions.