--- a/tools/integration_tests/util/setup/setup.go
+++ b/tools/integration_tests/util/setup/setup.go
@@ -261,6 +261,8 @@ func ParseSetUpFlags() {
 	}
 }
 
+// ParseSetUpFlagsForStretchrTests just parses the flags but doesn't fail until ParseSetUpFlags.
+// It skips the test gracefully if the integration tests are not being run.
 func ParseSetUpFlagsForStretchrTests(t *testing.T) {
 	flag.Parse()