GCSFuse metrics have been implemented in OTel. This means that one can now use the available OTel exporters to export these metrics.

Perf results:

| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 441.29MiB/s  | 1.35MiB/s  |  62.67MiB/s  |  1.17MiB/s   |
|   PR   |  0.25MiB   | 431.12MiB/s  | 1.33MiB/s  |  63.21MiB/s  |  1.29MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3753.6MiB/s  | 77.41MiB/s | 1198.09MiB/s |  81.6MiB/s   |
|   PR   | 48.828MiB  | 4216.18MiB/s | 81.05MiB/s | 1199.31MiB/s |  81.72MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4065.37MiB/s | 26.02MiB/s | 373.55MiB/s  |  30.35MiB/s  |
|   PR   | 976.562MiB | 4239.34MiB/s | 33.89MiB/s | 447.68MiB/s  |  31.36MiB/s  |
### Link to the issue in case of a bug fix.
b/380003794

### Testing details
1. Manual - Verified that the metrics are coming up in Prometheus exports
2. Unit tests - NA
3. Integration tests - Added integration tests to verify that the Prometheus exports are happening correctly.